requirements.txt
antlr4-python3-runtime
rich